requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
coverage==7.4.1
Django==4.2.4
django-filter==23.2
django-forms-dynamic==1.0.0
django-widget-tweaks==1.4.12
factory-boy==3.3.0
Faker==22.6.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
mock==5.1.0
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pytest==8.0.0
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.4
